+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.0.0] - 2021-06-06
+
+### Changed:
+- `sessionScope` is not a string
+
+### Removed:
+- Backward compatibility with cross domain localstorage
+
 ## [7.2.0] - 2021-06-05
 ### Added:
 - Allow specifying of `cookieDomain` in config to add interceptors to multiple API subdomain: https://github.com/supertokens/supertokens-website/issues/58